Namespace AutomationSpecHelper to Spec::Support::AutomationHelper
