+++ spec/lib/miq_automation_engine/miq_ae_assertion_spec.rb
@@ -1,5 +1,5 @@
 describe MiqAeEngine::MiqAeObject do
-  include AutomationSpecHelper
+  include Spec::Support::AutomationHelper
 
   context "Expression" do
     let(:user) { FactoryGirl.create(:user_with_group) }
