+++ spec/lib/miq_automation_engine/miq_ae_method_dispatch_spec.rb
@@ -1,5 +1,5 @@
 describe "MiqAeMethodDispatch" do
-  include AutomationSpecHelper
+  include Spec::Support::AutomationHelper
 
   before do
     @method_name     = 'MY_METHOD'
