+++ spec/lib/miq_automation_engine/models/miq_ae_datastore_spec.rb
@@ -1,5 +1,5 @@
 describe MiqAeDatastore do
-  include AutomationSpecHelper
+  include Spec::Support::AutomationHelper
 
   before(:each) do
     @ver_fname = File.expand_path(File.join(File.dirname(__FILE__), "version.xml"))
