+++ spec/presenters/tree_builder_ae_class_spec.rb
@@ -1,5 +1,5 @@
 describe TreeBuilderAeClass do
-  include AutomationSpecHelper
+  include Spec::Support::AutomationHelper
 
   context "initialize" do
     before do
