+++ spec/spec_helper.rb
@@ -53,7 +53,7 @@ RSpec.configure do |config|
   config.include UiConstants,          :type => :controller
   config.include AuthHelper,           :type => :controller
 
-  config.include AutomationSpecHelper,   :type => :automation
+  config.include Spec::Support::AutomationHelper, :type => :automation
   config.include AutomationExampleGroup, :type => :automation
   config.define_derived_metadata(:file_path => /spec\/automation/) do |metadata|
     metadata[:type] ||= :automation
